+++ store-common/src/test/java/store/common/json/ValueReadingAndWritingTest.java
@@ -22,7 +22,7 @@ import store.common.json.schema.Schema;
 import store.common.value.Value;
 import store.common.value.ValueType;
 import static store.common.value.ValueType.BOOLEAN;
-import static store.common.value.ValueType.LONG;
+import static store.common.value.ValueType.INTEGER;
 import static store.common.value.ValueType.STRING;
 
 /**
@@ -56,20 +56,20 @@ public class ValueReadingAndWritingTest {
         put(Value.of(new Date(1391878872)),
             jsonNumber(1391878872));
 
-        put(Value.of(ImmutableMap.of("num", VALUES.get(LONG),
+        put(Value.of(ImmutableMap.of("num", VALUES.get(INTEGER),
                                      "text", VALUES.get(STRING),
                                      "bool", VALUES.get(BOOLEAN))),
             createObjectBuilder()
-                .add("num", JSON.get(LONG))
+                .add("num", JSON.get(INTEGER))
                 .add("text", JSON.get(STRING))
                 .add("bool", JSON.get(BOOLEAN))
                 .build());
 
-        put(Value.of(asList(VALUES.get(LONG),
+        put(Value.of(asList(VALUES.get(INTEGER),
                             VALUES.get(STRING),
                             VALUES.get(BOOLEAN))),
             createArrayBuilder()
-                .add(JSON.get(LONG))
+                .add(JSON.get(INTEGER))
                 .add(JSON.get(STRING))
                 .add(JSON.get(BOOLEAN))
                 .build());
